requirements.txt
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
boto3==1.20.49
botocore==1.23.49
click==8.0.3
decorator==5.1.1
executing==0.8.2
ipdb==0.13.9
ipython==8.0.1
jedi==0.18.1
jmespath==0.10.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.0
prompt-toolkit==3.0.27
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.19.2
s3transfer==0.5.1
six==1.16.0
stack-data==0.1.4
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
urllib3==1.26.8
wcwidth==0.2.5
statsmodels==0.13.2